Replace Clean URI API with Spoo.me API for URL shortening functionality
