--- app/api/cleanuri/route.ts
@@ -1,32 +1,0 @@
-import { NextResponse } from "next/server";
-
-export async function POST(req: Request) {
-  try {
-    const { url } = await req.json();
-
-    if (!url) {
-      return NextResponse.json({ error: "URL is required" }, { status: 400 });
-    }
-
-    const response = await fetch("https://cleanuri.com/api/v1/shorten", {
-      method: "POST",
-      headers: { "Content-Type": "application/x-www-form-urlencoded" },
-      body: new URLSearchParams({ url }),
-    });
-
-    if (!response.ok) {
-      return NextResponse.json(
-        { error: "Failed to shorten URL" },
-        { status: 500 },
-      );
-    }
-
-    const data = await response.json();
-    return NextResponse.json(data, { status: 200 });
-  } catch {
-    return NextResponse.json(
-      { error: "Something went wrong" },
-      { status: 500 },
-    );
-  }
-}
